+++ src/main/java/com/Clubbr/Clubbr/Repository/stablishmentRepo.java
@@ -2,5 +2,5 @@ package com.Clubbr.Clubbr.Repository;
 
 import com.Clubbr.Clubbr.Entity.stablishment;
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface stablishmentRepo extends JpaRepository<stablishment, Integer>{
+public interface stablishmentRepo extends JpaRepository<stablishment, Long>{
 }
